requirement.txt
agecalc==3.5.2
appdirs==1.4.3
attr==0.2.0
attrs==17.1.0
autopep8==1.3.5
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.3
configparser==3.5.0
decorator==4.0.11
Django==1.9.12
django-ckeditor==5.2.1
django-cors-headers==1.2.2
django-cors-middleware==1.3.1
django-extensions==1.7.6
django-mutant==0.2.1
django-picklefield==1.0.0
django-polymodels==1.4.6
django-rest-swagger==0.3.8
django-silk==2.0.0
django-simple-history==1.8.2
djangorestframework==3.5.4
entrypoints==0.2.3
enum34==1.1.6
fiscalyear==0.1.0
functools32==3.2.3.post2
futures==3.2.0
gprof2dot==2016.10.13
html5lib==1.0.1
ipaddress==1.0.22
ipdb==0.10.3
ipykernel==4.8.2
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==7.3.0
Jinja2==2.10
jsonfield==2.0.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
mutant==0.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.14.1
olefile==0.44
packaging==16.8
pandocfilters==1.4.2
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
pkg-resources==0.0.0
prometheus-client==0.3.0
prompt-toolkit==1.0.14
psycopg2==2.7.5
ptyprocess==0.5.1
pycodestyle==2.4.0
pyexcel==0.4.5
pyexcel-io==0.3.4.1
Pygments==2.2.0
pyldap==2.4.28
pyparsing==2.2.0
python-box==2.2.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==17.1.0
qtconsole==4.3.1
requests==2.13.0
scandir==1.5
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.2.4
terminado==0.8.1
testpath==0.3.1
texttable==0.9.0
tornado==5.1
traitlets==4.3.2
unicodecsv==0.14.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.3.0
xlrd==1.0.0
XlsxWriter==0.9.8
